requirements.txt
requests
rich
luma.led-matrix
RPi.GPIO
spidev
python-dotenv
